requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
qdrant-client==1.9.1
google-generativeai==0.8.3
python-dotenv==1.0.1
numpy==2.1.2
psycopg[binary]==3.2.1
